Fix incorrect file-icon when displaying recent workspaces
Fixes #4949

Fixes minor issue where the file-icon displayed when displaying
`recent workspaces` is incorrect. The issue occurs when a folder
is opened with a `.` in the beginning (ex: .theia). In order to fix
the issue, a `FileStat` is used instead of the `URI` since it holds
more complete data on what the appropriate file-icon should be.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>